requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Django==3.1
django-bootstrap4==2.2.0
djangorestframework==3.11.1
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
urllib3==1.25.10
